HBX-1638: Move class 'org.hibernate.tool.hbm2x.ExporterSettings' to 'org.hibernate.tool.internal.export.ExporterSettings'
