+++ orm/src/main/java/org/hibernate/tool/internal/export/ExporterSettings.java
@@ -1,4 +1,4 @@
-package org.hibernate.tool.hbm2x;
+package org.hibernate.tool.internal.export;
 
 public interface ExporterSettings {
